trie: rename trie/types to trie/trienode
